feat: update tournament user skills avg upon registration

+++ database/migrations/2019_12_12_000002_create_tournaments_table.php
@@ -17,7 +17,7 @@ class CreateTournamentsTable extends Migration
             $table->increments('id');
             $table->string('title', 255);
             $table->string('description', 255);
-            $table->float('cost');
+            $table->unsignedInteger('cost');
             $table->unsignedInteger('capacity');
             $table->dateTime('start_date');
             $table->dateTime('end_date');
